Add new Dropdown Element | Redo Text Input Element | Fix Text Paragraph | New Button Element | Fix custom events being removed after calling unreg instead of unregistering

+++ src/main/kotlin/meowing/zen/feats/meowing/MeowCount.kt
@@ -44,7 +44,7 @@ object meowcount : Feature("meowcount") {
 @Zen.Command
 object MeowCommand : CommandUtils("meowcount", aliases = listOf("zenmeow", "zenmeowcount")) {
     override fun execute(context: CommandContext<FabricClientCommandSource>): Int {
-        val count = meowcount.getMeowCount()
+        val count = meowcount.getMeowCount().toInt()
         ChatUtils.addMessage("$prefix §fYou have meowed §b$count §ftimes!")
         return 1
     }
